Display cpu usage [##112866861]

+++ Rakefile
@@ -1,11 +1,31 @@
 task :default => [:tests]
 
 desc 'run acceptance tests'
-task :tests do
-  pid = spawn(".stack-work/dist/x86_64-osx/Cabal-1.22.5.0/build/stethoscope-exe/stethoscope-exe")
+task :tests => [:compile] do
+  status = system 'stack test'
+  raise "Backend tests failed" unless status
+
+  pid = spawn(".stack-work/dist/x86_64-osx/Cabal-1.22.5.0/build/stethoscope-exe/stethoscope-exe",
+              :out => 'spec/logs/server.out',
+              :err => "spec/logs/server.err")
   Process.detach(pid)
 
   puts `rspec`
 
   Process.kill('TERM', pid)
+end
+
+desc 'compile frontend'
+task :compile do
+  Dir.chdir 'frontend' do
+    puts 'compiling frontend...'
+    status = system 'elm make Main.elm --output=../public/index.html'
+    raise "Failed to compile frontend" unless status
+  end
+end
+
+desc 'run app'
+task :run => [:compile] do
+  puts 'running app...'
+  system 'stack build && stack exec stethoscope-exe'
 end
